Add a test for external MGS pool.
Add a new modifier to MakeTest() to create a pool of MGSs that can be
used as an external MGS. As part of this, the way storage allocations are
specified in the Servers resource was changed to have different behavior for
AllocatePerCompute, AllocateAcrossServers, and AllocateSingleServer. This change
affects all existing tests.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>