Added container-profile-support + container timeout tests + fixed global lustre
Added new tests to check for proper error handling in containers during
PreRun and PostRun.

To support this, WithContainerProfile() was added to take an existing
"base" profile and make changes to the timeouts and retryLimit to ease
testing.

Additionally, there were some issues with the global-lustre container
tests. Those are now fixed.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>