Fix slack-notify boolean for cron jobs
There was an issue where all 3 cases for slack-notify were not working
correctly:
  - cron jobs
  - manual trigger with slack-notify checked
  - manual trigger with slack-notify unchecked

This makes all 3 work as expected. There was some annoying behavior with
how github treats booleans and that was working against the environment
variable override logic.

This was fixed by changing the override logic to just look at the
trigger type vs trying to use non-null values.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>