system-tests: use `env` file to override default variables
This file is ignored by git but can be used to override the default env
vars (e.g. N, Q, R) that can be specific for a system. If the `env` file
is not there, use the defaults in the `Makefile`.

This is useful to default the system-test to use a particular queue or
compute nodes using the `Q` or `R` env vars.

An example is provided with `env.example`.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>

added env.example