system-test: Add capacity tests (#95)
This adds capacity test to ensure that 80% of the requested capacity is
available on the provided filesystem.

To support this, test scripts must be used to submit to flux alloc. A
temp directory is now created for each test to ensure a safe place to
put the test script. This directory must reside in a location is that
available on the compute nodes.

The root of the directory can be set with `TEST_TMPDIR_PREFIX` and
defaults to `${HOME}/nnf/tmp`.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>